+++ lib/tests/test_panel_model.py
@@ -1,17 +1,19 @@
 from lib.panel_model import Panel
 
+import numpy as np
+
 
 def test_solve_output():
-    efficiency = 1.0
-    area = 1.0
-    maximum_output_power = 100.0
+    efficiency = np.float64(1.0)
+    area = np.float64(1.0)
+    maximum_output_power = np.float64(100.0)
 
     panel = Panel(efficiency, area, maximum_output_power)
 
-    irradiation = 0.0
+    irradiation = np.float64(0.0)
     assert panel.solve_output(irradiation) == irradiation
 
-    irradiation = 1.0
+    irradiation = np.float64(1.0)
     assert panel.solve_output(irradiation) == irradiation
 
     irradiation = maximum_output_power * 10
